Moved empty string deserialization to separate GsonFactory class
