+++ bitmex-api/src/main/java/io/contek/invoker/bitmex/api/rest/user/GetPosition.java
@@ -3,6 +3,7 @@ package io.contek.invoker.bitmex.api.rest.user;
 import com.google.gson.Gson;
 import io.contek.invoker.bitmex.api.common._Position;
 import io.contek.invoker.bitmex.api.rest.user.GetPosition.Response;
+import io.contek.invoker.commons.GsonFactory;
 import io.contek.invoker.commons.actor.IActor;
 import io.contek.invoker.commons.rest.RestContext;
 import io.contek.invoker.commons.rest.RestMethod;
@@ -19,7 +20,7 @@ import static io.contek.invoker.commons.rest.RestMethod.GET;
 @NotThreadSafe
 public final class GetPosition extends UserRestRequest<Response> {
 
-  private static final Gson gson = new Gson();
+  private static final Gson gson = GsonFactory.buildGson();
 
   private Integer count;
   private final Map<String, String> filter = new HashMap<>();
