+++ bitstamp-api/src/main/java/io/contek/invoker/bitstamp/api/websocket/WebSocketMessageParser.java
@@ -6,6 +6,7 @@ import com.google.gson.JsonObject;
 import io.contek.invoker.bitstamp.api.websocket.common.WebSocketRequestConfirmationMessage;
 import io.contek.invoker.bitstamp.api.websocket.market.DiffOrderBookChannel;
 import io.contek.invoker.bitstamp.api.websocket.market.LiveTradesChannel;
+import io.contek.invoker.commons.GsonFactory;
 import io.contek.invoker.commons.websocket.AnyWebSocketMessage;
 import io.contek.invoker.commons.websocket.IWebSocketMessageParser;
 
@@ -18,7 +19,7 @@ import static io.contek.invoker.bitstamp.api.websocket.common.constants.WebSocke
 @Immutable
 final class WebSocketMessageParser implements IWebSocketMessageParser {
 
-  private final Gson gson = new Gson();
+  private final Gson gson = GsonFactory.buildGson();
 
   static WebSocketMessageParser getInstance() {
     return InstanceHolder.INSTANCE;
