+++ bybit-api/src/main/java/io/contek/invoker/bybit/api/websocket/WebSocketMessageParser.java
@@ -6,6 +6,7 @@ import com.google.gson.JsonObject;
 import io.contek.invoker.bybit.api.websocket.common.WebSocketOperationResponse;
 import io.contek.invoker.bybit.api.websocket.common.WebSocketTopicMessage;
 import io.contek.invoker.bybit.api.websocket.market.OrderBookChannel;
+import io.contek.invoker.commons.GsonFactory;
 import io.contek.invoker.commons.websocket.AnyWebSocketMessage;
 import io.contek.invoker.commons.websocket.IWebSocketComponent;
 import io.contek.invoker.commons.websocket.IWebSocketMessageParser;
@@ -20,7 +21,7 @@ import static io.contek.invoker.bybit.api.websocket.common.constants.WebSocketDa
 @Immutable
 final class WebSocketMessageParser implements IWebSocketMessageParser {
 
-  private final Gson gson = new Gson();
+  private final Gson gson = GsonFactory.buildGson();
 
   private final Map<String, Class<? extends WebSocketTopicMessage>> channelMessageTypes =
       new HashMap<>();
