+++ deribit-api/src/main/java/io/contek/invoker/deribit/api/websocket/WebSocketMessageParser.java
@@ -3,6 +3,7 @@ package io.contek.invoker.deribit.api.websocket;
 import com.google.gson.Gson;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
+import io.contek.invoker.commons.GsonFactory;
 import io.contek.invoker.commons.websocket.AnyWebSocketMessage;
 import io.contek.invoker.commons.websocket.IWebSocketMessageParser;
 import io.contek.invoker.deribit.api.websocket.common.WebSocketInboundMessage;
@@ -17,7 +18,7 @@ import javax.annotation.concurrent.Immutable;
 @Immutable
 final class WebSocketMessageParser implements IWebSocketMessageParser {
 
-  private final Gson gson = new Gson();
+  private final Gson gson = GsonFactory.buildGson();
 
   static WebSocketMessageParser getInstance() {
     return InstanceHolder.INSTANCE;
