+++ ftx-api/src/main/java/io/contek/invoker/ftx/api/rest/market/GetFutures.java
@@ -3,7 +3,7 @@ package io.contek.invoker.ftx.api.rest.market;
 import io.contek.invoker.commons.actor.IActor;
 import io.contek.invoker.commons.rest.RestContext;
 import io.contek.invoker.commons.rest.RestParams;
-import io.contek.invoker.ftx.api.common._Market;
+import io.contek.invoker.ftx.api.common._Future;
 import io.contek.invoker.ftx.api.rest.common.RestResponse;
 
 import javax.annotation.concurrent.NotThreadSafe;
@@ -32,5 +32,5 @@ public final class GetFutures extends MarketRestRequest<GetFutures.Response> {
   }
 
   @NotThreadSafe
-  public static final class Response extends RestResponse<List<_Market>> {}
+  public static final class Response extends RestResponse<List<_Future>> {}
 }
