+++ hbdminverse-api/src/main/java/io/contek/invoker/hbdminverse/api/websocket/common/marketdata/MarketDataWebSocketMessageParser.java
@@ -3,6 +3,7 @@ package io.contek.invoker.hbdminverse.api.websocket.common.marketdata;
 import com.google.gson.Gson;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
+import io.contek.invoker.commons.GsonFactory;
 import io.contek.invoker.commons.websocket.AnyWebSocketMessage;
 import io.contek.invoker.commons.websocket.IWebSocketComponent;
 import io.contek.invoker.commons.websocket.IWebSocketMessageParser;
@@ -20,7 +21,7 @@ import java.util.zip.GZIPInputStream;
 @ThreadSafe
 public final class MarketDataWebSocketMessageParser implements IWebSocketMessageParser {
 
-  private final Gson gson = new Gson();
+  private final Gson gson = GsonFactory.buildGson();
 
   private final Map<String, Class<? extends MarketDataWebSocketChannelMessage>>
       channelMessageTypes = new HashMap<>();
